requirements.txt
speech_recognition==3.8.1
pyaudio==0.2.11
requests==2.24.0
json==2.0.9
python-dotenv==0.13.0
